Add files via upload
Hi John,
I have excluded the documentation as it was overcrowding the repository and I dont want to take up too much of your time. Basketballplayer and Hockeyplayer classes are created to be used in BaskeyballPlayerTests  and HockeyPlayerTests respectively. I also included the test plan. There is also a screenshot of the requirements of the entire assignment should you want to take a look. Again this was a very time consuming assignment and I dont expect you to invest a lot of time into it. Maybe you can easily detect the problems who knows.